Eliminar dependencia y separar compilacion por despliegue

+++ script.sh
@@ -17,10 +17,14 @@ elif [ "$1" = "firebase" ]
 # utilizando Angular Github Pages
 	then
 		echo "Inicia despliegue con Firebase"
-		rm -rf dist
-		echo "Carpeta /dist borrada. Comienzo de la compilación"
-		ng build --prod=true --aot=true --buildOptimizer=true --optimization=true --vendorChunk=true
-		echo "Aplicación compilada. Lista para cargar al servidor"
+		if [ "$2" = "compile" ]
+			then
+				rm -rf dist
+				echo "Carpeta /dist borrada. Comienzo de la compilación"
+				ng build --prod=true --aot=true --buildOptimizer=true --optimization=true --vendorChunk=true
+				echo "Aplicación compilada"
+		fi
+		echo "Lista para cargar al servidor"
 		firebase deploy --only hosting:intercambios-sin-fronteras
 		echo "Aplicación desplegada"
 		echo "Puede visualizarla en https://intercambios-sin-fronteras.web.app"
